I completed user story two (Make a form which adds to the to-do list).
In index.html I created a form with two inputs so
the user can insert data.

In todos.py I created a query to insert data on to
the index.html

In todos.py I created a nested with statment so
the database had a connection and could complete the
query I previously created.